refactor(foundation): add PromotionService trait + drop allow(deprecated) from mockforge-pipelines
Phase 6 / Sub-phase A11: Add PromotionService trait to foundation matching
core's exact signature. Now that MockEnvironmentName + PromotionEntityType
are in foundation (A10), the trait can also live there.

Foundation:
- workspace_promotion::PromotionService trait (#[async_trait], 8-arg
  promote_entity matching mockforge-core's existing signature)

Core:
- workspace/promotion_trait.rs: re-exports PromotionService from foundation

Pipelines migration:
- src/steps/auto_promote.rs: imports PromotionService from foundation
- Removed #![allow(deprecated)] — 70+7 tests pass, clippy -D warnings clean
- Fixed pre-existing unused variable warnings in tests/integration_tests.rs
  (renamed to _executor / _notify_step)

Three consumer crates now fully unblocked:
- mockforge-vbr (A7) — 359 tests
- mockforge-chaos (A9) — 339 tests
- mockforge-pipelines (A11) — 70+7 tests
