+++ crates/mockforge-pipelines/src/lib.rs
@@ -1,6 +1,3 @@
-// TODO: remove once mockforge-workspace crate is fully extracted (needs PromotionService, PromotionEntityType, MockEnvironmentName)
-#![allow(deprecated)]
-
 //! # `MockForge` Pipelines
 //!
 //! Event-driven pipeline orchestration for `MockForge`.
